Add the sequence number in CalendarEvent and in CalendarEventOccurrence to represent a revision sequence number within a sequence of revisions. This sequence number follows as least the requirements of the icalendar specification about the sequence property of a calendar component. It is mainly dedicated for the synchronization and the syndications of events in external calendars

+++ core-library/src/integration-test/resources/org/silverpeas/core/calendar/create_table_calendar.sql
@@ -183,7 +183,7 @@ CREATE TABLE IF NOT EXISTS SB_Cal_Components (
   location       VARCHAR(255)  NULL,
   attributes     VARCHAR(40)   NULL,
   priority       INT           NOT NULL,
-  sequence       INT           NOT NULL DEFAULT 0,
+  sequence       INT8          NOT NULL DEFAULT 0,
   createDate     TIMESTAMP     NOT NULL,
   createdBy      VARCHAR(40)   NOT NULL,
   lastUpdateDate TIMESTAMP     NOT NULL,
